testing database connection added to index.js
